feat: moved to BondNominate
