+++ src/renderer/widgets/Staking/BondNominate/index.ts
