+++ src/renderer/widgets/Staking/BondNominate/lib/bond-utils.ts
