+++ src/renderer/widgets/Staking/BondNominate/lib/types.ts
