+++ src/renderer/widgets/Staking/BondNominate/model/bond-model.ts
