+++ src/renderer/widgets/Staking/BondNominate/model/confirm-model.ts
