+++ src/renderer/widgets/Staking/BondNominate/model/validators-model.ts
