+++ src/renderer/widgets/Staking/BondNominate/ui/Bond.tsx
