+++ src/renderer/widgets/Staking/BondNominate/ui/BondForm.tsx
