+++ src/renderer/widgets/Staking/BondNominate/ui/Confirmation.tsx
