+++ src/renderer/widgets/Staking/BondNominate/ui/Validators.tsx
